+++ src/mesh/uv_mapping.rs
@@ -143,14 +143,14 @@ impl UVOptions {
 
     /// Default values for hexagonal planes or column caps
     #[must_use]
-    #[deprecated(since = "0.13.1", note = "Use `UVOptions::new` instead")]
+    #[deprecated(since = "0.14.0", note = "Use `UVOptions::new` instead")]
     pub const fn cap_default() -> Self {
         Self::new()
     }
 
     /// Default values for quads
     #[must_use]
-    #[deprecated(since = "0.13.1", note = "Use `UVOptions::new` instead")]
+    #[deprecated(since = "0.14.0", note = "Use `UVOptions::new` instead")]
     pub const fn quad_default() -> Self {
         Self::new()
     }